+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Release 0.7.0 (in progress)
+
 ## Release 0.6.0 (2022-06-29)
 
 **Features**: